Reworked fingerprint generation and incremented version to v0.3.2-1.
