+++ src/consts.py
@@ -23,7 +23,7 @@ AUTH_PARAMS = {
     "window_title": "Login to Rockstar Games Social Club",
     "window_width": 700,
     "window_height": 600,
-    "start_uri": "http://127.0.0.1:8000",
+    "start_uri": "https://www.rockstargames.com/auth/scauth-login",
     "end_uri_regex": r"https://www.rockstargames.com/auth/get-user.json.*"
 }
