+++ src/manifest.json
@@ -2,7 +2,7 @@
     "name": "Rockstar Integration Plugin for Galaxy 2.0",
     "platform": "rockstar",
     "guid": "774732b5-69c4-405c-b6c9-92cd55740cfe",
-    "version": "0.3.2",
+    "version": "0.3.2-1",
     "description": "This plugin will allow for (basic) integration with the Rockstar Games Launcher.",
     "author": "Tylerbrawl",
     "email": "tylerbrawl@gmail.com",
