+++ src/version.py
@@ -1,1 +1,1 @@
-__version__ = "0.3.2"
+__version__ = "0.3.2-1"
